refactor(heartbeat): split heartbeat.ts into modules
heartbeat.ts was 701 lines covering state persistence, the cadence
scheduler, and the agent run. Split into a heartbeat/ folder:
  - state     — shared run-guard/timers/config holder + persisted-state I/O
  - agent     — prompt + goal-block building, the one-shot agent run with
                timeout/abort/orphan-eviction, log helpers
  - scheduler — init, cadence timer (scheduler-core driven), run guard,
                force/await/status API

index.ts re-exports the identical public surface (incl. the test-only
buildHeartbeatSystemPrompt + renderGoalsBlock). Importers updated to the
barrel. Verified: heartbeat suite passes (36 tests).
